pyproject.toml
[tool.poetry]
name = "speaker-diarization"
version = "0.1.0"
description = ""
authors = ["Jared Wuliger <jwuliger.dev@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
google-cloud-speech = "^2.26.1"
google-cloud-storage = "^2.18.0"
pydub = "^0.25.1"
black = "^24.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
